Build in site/dist, remove unneeded assets
After this change, everything needed to run hugworm is contained in
site/, and nothing else is present (except old builds).

It's now easier to publish somewhere online and harder to forget the
`/site` in the URL :)

I'll admit I don't know what the package-lock.json file was doing in
site. It doesn't seem like anything depend on it.